+++ src/Makefile.gtest.include
@@ -9,6 +9,8 @@ noinst_PROGRAMS += zcash-gtest
 # test_checktransaction.cpp MUST be before
 # any test that calls SelectParams().
 zcash_gtest_SOURCES = \
+	test/testutil.cpp \
+	test/testutil.h \
 	gtest/main.cpp \
 	gtest/utils.cpp \
 	gtest/utils.h \
