+++ src/Makefile.test.include
@@ -102,6 +102,8 @@ BITCOIN_TESTS =\
   test/test_random.h \
   test/test_util.cpp \
   test/test_util.h \
+  test/testutil.cpp \
+  test/testutil.h \
   test/torcontrol_tests.cpp \
   test/transaction_tests.cpp \
   test/txvalidationcache_tests.cpp \
