+++ src/test/test_bitcoin.cpp
@@ -25,6 +25,8 @@
 #include "rpc/server.h"
 #include "rpc/register.h"
 
+#include "test/testutil.h"
+
 #include <boost/filesystem.hpp>
 #include <boost/test/unit_test.hpp>
 #include <boost/thread.hpp>
